Add a note preventing wrong AULA_SAMPLES setting; vim use seems mandatory
If AULA_SAMPLES is inside any directory bound the docker,
click-dummies-refresh always refreshes all files. Even changing
the timezone to match that of the host machine doesn't help.
Unfortunately, that means the directory can't easily be edited
from outside of the docker (changing permissions
on /var/lib/docker/aufs/mnt/... doesn't seem to help).
If vim is not enough for everyone, nano can be copied to ~/aula/
inside the docker.