Fix: Frames, Pages, Forms.
Before, redirectFormHandler called toHtml when rendering form pages on
the already-rendered page in a not-yet-rendered Frame.

This was in conflict with the refactorings from earlier today that
moved the extraPageHeaders in the Page class.

Now, we only collect all information needed for form page rendering
inside redirectFormHandler, but make servant look up the ToHtml
instance and trigger the actual rendering.