DEBUGGING.
Somebody is illegally calling listIdeas' with class space and topic tab.
I commit this becasue it exposes an interesting space leak: the error in
listIdeas' is thrown when testing with `cabal test`, but runs into a mem
leak when testing with `make sensei` (careful!  use `ulimit -d` to
reproduce that!).

I think this may be a bug in our test suite, something about cleaning up
resources that is not working properly, but only triggered when re-running
tests in a loop (like sensei does).

I am committing this for the record, in the hope we'll find it again if
we have to solve this issue in the future.  for now, i will focus on
finding the bug in listIdeas'.