Fix: profile edit.
We will need permissions to prevent one student to edit another's
profile, but that's not the point here.  The question whether one
user (e.g. a moderator or admin) should be able to edit another
user's profile (e.g., because there is offensive content in the
image, but not in the description), is one of authorization.