Remove outdated comment
Now we have the IdeaSpace in the context. In the example above a student of class 2016-6a,
can visit /space/2016-7a/ideas but the list of ideas is empty and then trying to add an idea
shows the access denied. Since we don't have any way to reach these other classes the situation
this is not an issue.