docs.
i'm not sure what the best practice here is, so i only did a source
code comment.  but i was not convinced that this test was very clear
without this text.  i think i only understood the purpose because we
talked about it yesterday.  should this be a very long hspec
description instead?