Hack jury phase circumvention into the existing system.
this is a bit brutal, but it may work (needs testing!).

the idea is that we introduce an "action" that is triggered
if the jury phase is entered, and the action consists in
essentially pushing the "shift phase" button in the admin panel.

this is a minimal change, while it seems plausible to me that
it should work, as it makes maximum use of the existing
application logic.